src/api/routes/search.php src/EcclesiaCRM/APIControllers/SearchController.php
